[WIP] Faster subtrees for MiqReportResults
Avoids loading all the columns from `MiqReportResult` and associated
`MiqTask` objects, and simply loads the needed columns for the
presenter/decorator/tree.

Creates a decent amount of frozen strings, constants as well to achieve
this, as well as custom arel.  Some of which could have been done in the
core repo, but I skipped for now to make this an isolated change.