[FIXUP] Update cypress.rake
- Change namespace to `cypress:ui`
- Add `yarn_cmd` to allow running from manageiq dir

Made the switch with the namespec to avoid collisions with
`rspec-rails`'s rake task.  This only presented itself when running from
`manageiq`.  The namespace change also allows running the specs from
manageiq, but makes it so that it is "future proof" so when we add
cypress to other providers, they can follow the namespace conventions
here.

The `yarn_cmd` is a simple helper that allows setting the `--cwd` option
in yarn so it can work from the directory that contains the
`package.yml`.  Might need to be expanded in the future to work with
other repos, but this is a start for now.

This all might turn into a `Rake::Task` library instead of using the DSL
directly so that we can re-use this logic elsewhere.