[REMOVEME] "Fix" for menuItems command
Will remove from this branch, but by using a `cy.get` with this command,
it forces a `wait` for the DOM to load with the menu loaded.  Using
`cy.window` doesn't do that.

This probably could be smaller, but this is a POC commit to fix the
currently breaking tests using this method (currently only
`menu.spec.js`).